Verify existing cookies are authenticated
A runtime error cropped up that turned out to simply be that an
previously dumped cookies file was not authenticating me to DeviantArt.
This was then causing runtime problems against ristricted content.

auth() now verifies that existing cookies do provided authentication.

This however doesn't fix the problem as to why the program had a runtime
error when it encountered restricted content. I certainly don't want it
to fail for unauthenticated users; restricted content should just be
skipped.