Add method to download description for Deviations
Note that pretty_print=True seems to make no difference to the HTML
output. It's all to do with whitespace and preserving text data, and
since the description is basically one big text block inside a div
(obviously containing other elements like links and img's) it decideds
not to change anything.

You could replace spaces with newlines `\n` as HTML treats them all as
spaces, but lxml seems not to wrap etc or anything like that.

In the end this is a side issue; yes it would be nice to have a pretty
looking html document, but it all renders that same and I don't see any
situation where I would need to look at, let alone edit the source html
of the descriptions. So save fixing that for the rainy day when you have
no more projects to complete ;D