Identify restricted content
Restricted content is now identified when the session is not
authenticated against DeviantArt (eg because the user didn't or couldn't
log in, or the cookies became invalid etc). Such content is then
processed using the Deviation class.

This is important as previously restricted content would most likely be
identified as a named type (as opposed to `unknown`), however many of
the extra xpath queries made by the sub-classes associated with the
identified type would be invalid (the elements don't exist for
restricted content) and the program would crash.