Change working directory in place of passing path
Instead of passing a path to Deviation.downloadXXX(), let the collection
change the working directory. I am unsure if this would be considered
bad form, but to me it makes more sense. I will preserve the `path`
parameter on the .downloadXXX's however as it may be usefull in the
future (maybe :D).

The creation and change to the root encapsulating directory (ie the
username passed to dadl as an argument) was moved from dadl to
Favourites. Again that makes more sense to me.

Avatars are also now downloaded to the username/collection/avatar and
not username/avatar. The motivation for the original location was to do
with not repeating avatars accross different collections, but thats
really not an issue, they take up such a tiny amount of space. Plus its
probably more usefull to have all resources for a collection bundled
together. It is better :D