Remove partial files for remote connection errors
- If the program gets a connection error while making a request to the
  deviation, deviations description or the creators avatar, the empty
  file created to save that data locally is deleted.
- If the program gets a connection error while streaming data from a
  previously successful requests the empty file will be deleted.

Remote connections by far are the most runtime consuming part of the
program, so it is right that the program doesn't make them if it doesn't
have to. This means checking for local versions of the file before
making the requests, and as a result creating an empty file if it
doesn't exist before the remote connection is made.

This change makes it much easier to re-run the program to complete a
download of a favourites set if a connection error occured during a
previous run. Previously you would have to track down any empty or
partially complete files and manually delete them, as to the program
they would be considered complete files and their download would be
skipped.