Download the original image if it exists
Add a new field Deviation.downloadurl to hold the original files url.
For Data deviations this is just the same as Data.dataurl, however for
img deviations this may be different, and may point to a higher
resolution img. As such I have only made use of this feature for img
deviations.

In order to download the original image the options -d and -f must be
enabled. -f is obvious :D. -d is because the link to the orignal file
url exists only on the deviations page, and for most deviations the only
time when their page is loaded is to download the description.

In principle its perfectly possible to check for an original file url
for all img deviations, however that would mean loading every img
deviations page, even if the -d option wasn't given, which seems a waste
since you would then have the description in memory, yet you did nothing
with it.

I'm heading down the path of paradox of choice :D For now this is how it
works.