Add improved identification of restricted content
Implement a new way to check for restricted content that doesn't require
loading the deviations page, speeding up the program.

The old restricted content identification block has been left in place
for not (identified by the 'Legacy' tag). This is there if there are
other case's of restricted content that I have yet to encounter that are
not caught by this new check.

Catching restricted content at the begining of the id method was
motivated by an issue encountered in chesterocampo's favourites library.
An `antisocial` (ie restricted by the creator) `film` deviation was
identified as a `film` deviation because the xpath query to id films
works weather it is restricted or not. However this was running an
un-authenticated session, and the program subsequently crashed when it
tried to find the film file's url, which didn't exist on the deviations
page for a non-logged in user.

This new version has been tested against chesterocampo's favourites
library for both a logged in and non-logged in run of the program, and
both work as expected, sussessfully handling the restricted content for
the non run.